Some cleanup and formatting

+++ trunk/powercraft_16/forge/mcp/src/minecraft/powercraft/core/blocks/PC_BlockGenerator.java
@@ -1,7 +1,6 @@
 package powercraft.core.blocks;
 
-import cpw.mods.fml.relauncher.Side;
-import cpw.mods.fml.relauncher.SideOnly;
+
 import net.minecraft.block.material.Material;
 import net.minecraft.creativetab.CreativeTabs;
 import net.minecraft.entity.EntityLivingBase;
@@ -15,21 +14,27 @@ import powercraft.api.PC_Direction;
 import powercraft.api.PC_Utils;
 import powercraft.api.blocks.PC_Block;
 import powercraft.api.blocks.PC_BlockInfo;
-import powercraft.api.blocks.PC_TileEntity;
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 
-@PC_BlockInfo(name="Generator", blockid="generator", defaultid=3001, tileEntity = PC_TileEntityGenerator.class)
+
+@PC_BlockInfo(name = "Generator", blockid = "generator", defaultid = 3001, tileEntity = PC_TileEntityGenerator.class)
 public class PC_BlockGenerator extends PC_Block {
 
 	private Icon frontLevel[] = new Icon[4];
-	
+
+
 	public PC_BlockGenerator(int id) {
+
 		super(id, Material.ground);
 		setCreativeTab(CreativeTabs.tabDecorations);
 	}
 
+
 	@Override
 	@SideOnly(Side.CLIENT)
 	public void loadIcons() {
+
 		blockIcon = PC_ClientRegistry.registerIcon("DefaultMaschineTexture");
 		frontLevel[0] = PC_ClientRegistry.registerIcon("Front_Lvl0");
 		frontLevel[1] = PC_ClientRegistry.registerIcon("Front_Lvl1");
@@ -37,38 +42,44 @@ public class PC_BlockGenerator extends PC_Block {
 		frontLevel[3] = PC_ClientRegistry.registerIcon("Front_Lvl3");
 	}
 
+
 	@Override
 	public void registerRecipes() {
-		
+
 	}
-	
+
+
 	@Override
 	public void onBlockPlacedBy(World world, int x, int y, int z, EntityLivingBase entity, ItemStack itemStack) {
-		int l = MathHelper.floor_double((double)(entity.rotationYaw * 4.0F / 360.0F) + 0.5D) & 3;
+
+		int l = MathHelper.floor_double(entity.rotationYaw * 4.0F / 360.0F + 0.5D) & 3;
 		PC_Utils.setMD(world, x, y, z, PC_Direction.PLAYER2MD[l]);
 	}
+
 
 	@Override
 	@SideOnly(Side.CLIENT)
 	public Icon getBlockTexture(IBlockAccess world, int x, int y, int z, int side) {
-		if(side==PC_Utils.getRotation(world, x, y, z)){
+
+		if (side == PC_Utils.getRotation(world, x, y, z)) {
 			PC_TileEntityGenerator generator = PC_Utils.getTE(world, x, y, z);
-			if(generator!=null && generator.getHeat()>0){
-				return frontLevel[(generator.getHeat()-1)*3/PC_TileEntityGenerator.maxHeat+1];
-			}else{
+			if (generator != null && generator.getHeat() > 0) {
+				return frontLevel[(generator.getHeat() - 1) * 3 / PC_TileEntityGenerator.maxHeat + 1];
+			} else {
 				return frontLevel[0];
 			}
-		}else{
+		} else {
 			return blockIcon;
 		}
 	}
 
+
 	@Override
 	@SideOnly(Side.CLIENT)
 	public Icon getIcon(int side, int metadata) {
-		if(side==5)
-			return frontLevel[0];
+
+		if (side == 5) return frontLevel[0];
 		return blockIcon;
 	}
-	
+
 }
